add require js and fix dependencies. #20
