+++ www/js/core/lib/paramiko.js
@@ -1,4 +1,6 @@
 // Source: common.js
+define([], function () {
+
 function paramikojs() { }
 paramikojs = {
   MSG_DISCONNECT : 1,
@@ -212,27 +214,27 @@ sjcl.cipher.aes = function (key, mode) {
   if (keyLen !== 4 && keyLen !== 6 && keyLen !== 8) {
     throw "invalid aes key size";
   }
-  
+
   this._key = [encKey = key.slice(0), decKey = []];
-  
+
   // schedule encryption keys
   for (i = keyLen; i < 4 * keyLen + 28; i++) {
     tmp = encKey[i-1];
-    
+
     // apply sbox
     if (i%keyLen === 0 || (keyLen === 8 && i%keyLen === 4)) {
       tmp = sbox[tmp>>>24]<<24 ^ sbox[tmp>>16&255]<<16 ^ sbox[tmp>>8&255]<<8 ^ sbox[tmp&255];
-      
+
       // shift rows and add rcon
       if (i%keyLen === 0) {
         tmp = tmp<<8 ^ tmp>>>24 ^ rcon<<24;
         rcon = rcon<<1 ^ (rcon>>7)*283;
       }
     }
-    
+
     encKey[i] = encKey[i-keyLen] ^ tmp;
   }
-  
+
   // schedule decryption keys
   for (j = 0; i; j++, i--) {
     tmp = encKey[j&3 ? i : i - 4];
@@ -286,7 +288,7 @@ sjcl.cipher.aes.prototype = {
 
     return ret;
   },
-  
+
   /**
    * Decrypt an array of 4 big-endian words.
    * @param {Array} data The ciphertext.
@@ -308,7 +310,7 @@ sjcl.cipher.aes.prototype = {
     }
     return ret;
   },
-  
+
   /**
    * The expanded S-box and inverse S-box tables.  These will be computed
    * on the client so that we don't have to send them down the wire.
@@ -337,32 +339,32 @@ sjcl.cipher.aes.prototype = {
    for (i = 0; i < 256; i++) {
      th[( d[i] = i<<1 ^ (i>>7)*283 )^i]=i;
    }
-   
+
    for (x = xInv = 0; !sbox[x]; x ^= x2 || 1, xInv = th[xInv] || 1) {
      // Compute sbox
      s = xInv ^ xInv<<1 ^ xInv<<2 ^ xInv<<3 ^ xInv<<4;
      s = s>>8 ^ s&255 ^ 99;
      sbox[x] = s;
      sboxInv[s] = x;
-     
+
      // Compute MixColumns
      x8 = d[x4 = d[x2 = d[x]]];
      tDec = x8*0x1010101 ^ x4*0x10001 ^ x2*0x101 ^ x*0x1010100;
      tEnc = d[s]*0x101 ^ s*0x1010100;
-     
+
      for (i = 0; i < 4; i++) {
        encTable[i][x] = tEnc = tEnc<<24 ^ tEnc>>>8;
        decTable[i][s] = tDec = tDec<<24 ^ tDec>>>8;
      }
    }
-   
+
    // Compactify.  Considerable speedup on Firefox.
    for (i = 0; i < 5; i++) {
      encTable[i] = encTable[i].slice(0);
      decTable[i] = decTable[i].slice(0);
    }
   },
-  
+
   /**
    * Encryption and decryption core.
    * @param {Array} input Four words to be encrypted or decrypted.
@@ -374,7 +376,7 @@ sjcl.cipher.aes.prototype = {
     if (input.length !== 4) {
       throw "invalid aes block size";
     }
-    
+
     var key = this._key[dir],
         // state variables a,b,c,d are loaded with pre-whitened data
         a = input[0]           ^ key[0],
@@ -382,20 +384,20 @@ sjcl.cipher.aes.prototype = {
         c = input[2]           ^ key[2],
         d = input[dir ? 1 : 3] ^ key[3],
         a2, b2, c2,
-        
+
         nInnerRounds = key.length/4 - 2,
         i,
         kIndex = 4,
         out = [0,0,0,0],
         table = this._tables[dir],
-        
+
         // load up the tables
         t0    = table[0],
         t1    = table[1],
         t2    = table[2],
         t3    = table[3],
         sbox  = table[4];
- 
+
     // Inner rounds.  Cribbed from OpenSSL.
     for (i = 0; i < nInnerRounds; i++) {
       a2 = t0[a>>>24] ^ t1[b>>16 & 255] ^ t2[c>>8 & 255] ^ t3[d & 255] ^ key[kIndex];
@@ -405,18 +407,18 @@ sjcl.cipher.aes.prototype = {
       kIndex += 4;
       a=a2; b=b2; c=c2;
     }
-        
+
     // Last round.
     for (i = 0; i < 4; i++) {
       out[dir ? 3&-i : i] =
-        sbox[a>>>24      ]<<24 ^ 
+        sbox[a>>>24      ]<<24 ^
         sbox[b>>16  & 255]<<16 ^
         sbox[c>>8   & 255]<<8  ^
         sbox[d      & 255]     ^
         key[kIndex++];
       a2=a; a=b; b=c; c=d; d=a2;
     }
-    
+
     return out;
   }
 };
@@ -997,16 +999,16 @@ function des (key, message, encrypt, mode, iv, padding, keys) {
     temp = ((right >>> 8) ^ left) & 0x00ff00ff; left ^= temp; right ^= (temp << 8);
     temp = ((left >>> 1) ^ right) & 0x55555555; right ^= temp; left ^= (temp << 1);
 
-    left = ((left << 1) | (left >>> 31)); 
-    right = ((right << 1) | (right >>> 31)); 
+    left = ((left << 1) | (left >>> 31));
+    right = ((right << 1) | (right >>> 31));
 
     //do this either 1 or 3 times for each chunk of the message
     for (j=0; j<iterations; j+=3) {
       endloop = looping[j+1];
       loopinc = looping[j+2];
-      //now go through and perform the encryption or decryption  
+      //now go through and perform the encryption or decryption
       for (i=looping[j]; i!=endloop; i+=loopinc) { //for efficiency
-        right1 = right ^ keys[i]; 
+        right1 = right ^ keys[i];
         right2 = ((right >>> 4) | (right << 28)) ^ keys[i+1];
         //the result is attained by passing these bytes through the S selection functions
         temp = left;
@@ -1020,8 +1022,8 @@ function des (key, message, encrypt, mode, iv, padding, keys) {
     } //for either 1 or 3 iterations
 
     //move then each one bit to the right
-    left = ((left >>> 1) | (left << 31)); 
-    right = ((right >>> 1) | (right << 31)); 
+    left = ((left >>> 1) | (left << 31));
+    right = ((right >>> 1) | (right << 31));
 
     //now perform IP-1, which is IP in the opposite direction
     temp = ((left >>> 1) ^ right) & 0x55555555; right ^= temp; left ^= (temp << 1);
@@ -1100,7 +1102,7 @@ function des_createKeys (key) {
 
       //now apply PC-2, in such a way that E is easier when encrypting or decrypting
       //this conversion will look like PC-2 except only the last 6 bits of each byte are used
-      //rather than 48 consecutive bits and the order of lines will be according to 
+      //rather than 48 consecutive bits and the order of lines will be according to
       //how the S selection functions will be applied: S2, S4, S6, S8, S1, S3, S5, S7
       lefttemp = pc2bytes0[left >>> 28] | pc2bytes1[(left >>> 24) & 0xf]
               | pc2bytes2[(left >>> 20) & 0xf] | pc2bytes3[(left >>> 16) & 0xf]
@@ -1110,7 +1112,7 @@ function des_createKeys (key) {
                 | pc2bytes9[(right >>> 20) & 0xf] | pc2bytes10[(right >>> 16) & 0xf]
                 | pc2bytes11[(right >>> 12) & 0xf] | pc2bytes12[(right >>> 8) & 0xf]
                 | pc2bytes13[(right >>> 4) & 0xf];
-      temp = ((righttemp >>> 16) ^ lefttemp) & 0x0000ffff; 
+      temp = ((righttemp >>> 16) ^ lefttemp) & 0x0000ffff;
       keys[n++] = lefttemp ^ temp; keys[n++] = righttemp ^ (temp << 16);
     }
   } //for each iterations
@@ -1325,7 +1327,7 @@ kryptos.hash.baseHash.prototype = {
     for (var x = 0; x < this.data.length; ++x) {
       hashData.push(this.data.charCodeAt(x));
     }
-    
+
     if(!(window.Components && window.Components.classes)) {
       throw new Error("Unable to use " + this.type + " hash without Mozilla's Components.classes"); //FIXME
     }
@@ -1687,7 +1689,7 @@ kryptos.hash.MD5.prototype = {
     var keyObject = window.Components.classes["@mozilla.org/security/keyobjectfactory;1"]
                       .getService(window.Components.interfaces.nsIKeyObjectFactory)
                       .keyFromString(window.Components.interfaces.nsIKeyObject.HMAC, key);
-    
+
     hasher.init(digestmod, keyObject);
     var data = kryptos.toByteArray(msg);
     hasher.update(data, data.length);
@@ -1734,7 +1736,7 @@ kryptos.hash.MD5.prototype = {
 
 // Source: kryptos/PublicKey/RSA.js
 kryptos.publicKey.RSA = function() {
-  
+
 };
 
 kryptos.publicKey.RSA.prototype = {
@@ -1764,7 +1766,7 @@ kryptos.publicKey.RSA.prototype = {
 
 // Source: kryptos/PublicKey/DSA.js
 kryptos.publicKey.DSA = function() {
-  
+
 };
 
 kryptos.publicKey.DSA.prototype = {
@@ -2122,7 +2124,7 @@ kryptos.random.Fortuna.FortunaGenerator.AESGenerator.prototype = {
 
 // Source: kryptos/Random/OSRNG/browser.js
 kryptos.random.OSRNG.BrowserRNG = function() {
-  
+
 };
 
 kryptos.random.OSRNG.BrowserRNG.prototype = {
@@ -2154,7 +2156,7 @@ var binascii = {
     padding = padding || '';
     for (var x = 0; x < str.length; ++x) {
       var c = str.charCodeAt(x).toString(16);
-      result += (c.length == 1 ? '0' : '') + c + padding;      
+      result += (c.length == 1 ? '0' : '') + c + padding;
     }
     return result;
   },
@@ -2163,7 +2165,7 @@ var binascii = {
     var result = "";
     for (var x = 0; x < str.length; x += 2) {
       result += String.fromCharCode(parseInt(str.charAt(x) + str.charAt(x + 1), 16));
-    }  
+    }
     return result;
   }
 };
@@ -2275,9 +2277,9 @@ This software is covered under the following copyright:
  * The above copyright notice and this permission notice shall be
  * included in all copies or substantial portions of the Software.
  *
- * THE SOFTWARE IS PROVIDED "AS-IS" AND WITHOUT WARRANTY OF ANY KIND, 
- * EXPRESS, IMPLIED OR OTHERWISE, INCLUDING WITHOUT LIMITATION, ANY 
- * WARRANTY OF MERCHANTABILITY OR FITNESS FOR A PARTICULAR PURPOSE.  
+ * THE SOFTWARE IS PROVIDED "AS-IS" AND WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS, IMPLIED OR OTHERWISE, INCLUDING WITHOUT LIMITATION, ANY
+ * WARRANTY OF MERCHANTABILITY OR FITNESS FOR A PARTICULAR PURPOSE.
  *
  * IN NO EVENT SHALL TOM WU BE LIABLE FOR ANY SPECIAL, INCIDENTAL,
  * INDIRECT OR CONSEQUENTIAL DAMAGES OF ANY KIND, OR ANY DAMAGES WHATSOEVER
@@ -3518,7 +3520,7 @@ BigInteger = BigInteger.BigInteger;
   local machine.  If an SSH agent is running, this class can be used to
   connect to it and retreive L{PKey} objects which can be used when
   attempting to authenticate to remote SSH servers.
-  
+
   Because the SSH agent protocol uses environment variables and unix-domain
   sockets, this probably doesn't work on Windows.  It does work on most
   posix platforms though (Linux and MacOS X, for example).
@@ -3528,7 +3530,7 @@ paramikojs.Agent = function () {
     Open a session with the local machine's SSH agent, if one is running.
     If no agent is running, initialization will succeed, but L{get_keys}
     will return an empty tuple.
-    
+
     @raise SSHException: if an SSH agent is found, but speaks an
         incompatible protocol
   */
@@ -3561,7 +3563,7 @@ paramikojs.Agent = function () {
     // no agent support
     return;
   }
-      
+
   var msg = this._send_message(String.fromCharCode(paramikojs.Agent.SSH2_AGENTC_REQUEST_IDENTITIES));
   if (msg.ptype != paramikojs.Agent.SSH2_AGENT_IDENTITIES_ANSWER) {
     throw new paramikojs.ssh_exception.SSHException('could not get keys from ssh-agent');
@@ -3596,7 +3598,7 @@ paramikojs.Agent.prototype = {
     Return the list of keys available through the SSH agent, if any.  If
     no SSH agent was running (or it couldn't be contacted), an empty list
     will be returned.
-    
+
     @return: a list of keys available on the SSH agent
     @rtype: tuple of L{AgentKey}
   */
@@ -4088,7 +4090,7 @@ paramikojs.BER.prototype = {
   decode : function() {
     return this.decode_next();
   },
-    
+
   decode_next : function() {
     if (this.idx >= this.content.length) {
       return null;
@@ -4264,7 +4266,7 @@ paramikojs.Channel.prototype = {
     @type width: int
     @param height: height (in characters) of the terminal screen
     @type height: int
-    
+
     @raise SSHException: if the request was rejected or the channel was
         closed
   */
@@ -4290,14 +4292,14 @@ paramikojs.Channel.prototype = {
     Request an interactive shell session on this channel.  If the server
     allows it, the channel will then be directly connected to the stdin,
     stdout, and stderr of the shell.
-    
+
     Normally you would call L{get_pty} before this, in which case the
     shell will operate through the pty, and the channel will be connected
     to the stdin and stdout of the pty.
-    
+
     When the shell exits, the channel will be closed and can't be reused.
     You must open a new channel if you wish to open another shell.
-    
+
     @raise SSHException: if the request was rejected or the channel was
         closed
   */
@@ -4317,7 +4319,7 @@ paramikojs.Channel.prototype = {
     Execute a command on the server.  If the server allows it, the channel
     will then be directly connected to the stdin, stdout, and stderr of
     the command being executed.
-    
+
     When the command finishes executing, the channel will be closed and
     can't be reused.  You must open a new channel if you wish to execute
     another command.
@@ -4345,7 +4347,7 @@ paramikojs.Channel.prototype = {
     Request a subsystem on the server (for example, C{sftp}).  If the
     server allows it, the channel will then be directly connected to the
     requested subsystem.
-    
+
     When the subsystem finishes, the channel will be closed and can't be
     reused.
 
@@ -4400,7 +4402,7 @@ paramikojs.Channel.prototype = {
     status. You may use this to poll the process status if you don't
     want to block in L{recv_exit_status}. Note that the server may not
     return an exit status in some cases (like bad servers).
-    
+
     @return: True if L{recv_exit_status} will return immediately
     @rtype: bool
     @since: 1.7.3
@@ -4415,10 +4417,10 @@ paramikojs.Channel.prototype = {
     If the command hasn't finished yet, this method will wait until
     it does, or until the channel is closed.  If no exit status is
     provided by the server, -1 is returned.
-    
+
     @return: the exit code of the process on the server.
     @rtype: int
-    
+
     @since: 1.2
   */
   recv_exit_status : function() {
@@ -4430,10 +4432,10 @@ paramikojs.Channel.prototype = {
     really only makes sense in server mode.)  Many clients expect to
     get some sort of status code back from an executed command after
     it completes.
-    
+
     @param status: the exit code of the process
     @type status: int
-    
+
     @since: 1.2
   */
   send_exit_status : function(status) {
@@ -4452,25 +4454,25 @@ paramikojs.Channel.prototype = {
     Request an x11 session on this channel.  If the server allows it,
     further x11 requests can be made from the server to the client,
     when an x11 application is run in a shell session.
-    
+
     From RFC4254::
 
         It is RECOMMENDED that the 'x11 authentication cookie' that is
         sent be a fake, random cookie, and that the cookie be checked and
         replaced by the real cookie when a connection request is received.
-    
+
     If you omit the auth_cookie, a new secure random 128-bit value will be
     generated, used, and returned.  You will need to use this value to
     verify incoming x11 requests and replace them with the actual local
     x11 cookie (which requires some knoweldge of the x11 protocol).
-    
+
     If a handler is passed in, the handler is called from another thread
     whenever a new x11 connection arrives.  The default handler queues up
     incoming x11 connections, which may be retrieved using
     L{Transport.accept}.  The handler's calling signature is::
-    
+
         handler(channel: Channel, (address: str, port: int))
-    
+
     @param screen_number: the x11 screen number (0, 10, etc)
     @type screen_number: int
     @param auth_protocol: the name of the X11 authentication method used;
@@ -4561,21 +4563,21 @@ paramikojs.Channel.prototype = {
     Set whether stderr should be combined into stdout on this channel.
     The default is C{False}, but in some cases it may be convenient to
     have both streams combined.
-    
+
     If this is C{False}, and L{exec_command} is called (or C{invoke_shell}
     with no pty), output to stderr will not show up through the L{recv}
     and L{recv_ready} calls.  You will have to use L{recv_stderr} and
     L{recv_stderr_ready} to get stderr output.
-    
+
     If this is C{True}, data will never show up via L{recv_stderr} or
     L{recv_stderr_ready}.
-    
+
     @param combine: C{True} if stderr output should be combined into
         stdout on this channel.
     @type combine: bool
     @return: previous setting.
     @rtype: bool
-    
+
     @since: 1.1
   */
   set_combine_stderr : function(combine) {
@@ -4699,7 +4701,7 @@ paramikojs.Channel.prototype = {
     Returns true if data is buffered and ready to be read from this
     channel.  A C{False} result does not mean that the channel has closed;
     it means you may need to wait before more data arrives.
-    
+
     @return: C{True} if a L{recv} call on this channel would immediately
         return at least one byte; C{False} otherwise.
     @rtype: boolean
@@ -4718,7 +4720,7 @@ paramikojs.Channel.prototype = {
     @type nbytes: int
     @return: data.
     @rtype: str
-    
+
     @raise socket.timeout: if no data is ready before the timeout set by
         L{settimeout}.
   */
@@ -4747,11 +4749,11 @@ paramikojs.Channel.prototype = {
     channel's stderr stream.  Only channels using L{exec_command} or
     L{invoke_shell} without a pty will ever have data on the stderr
     stream.
-    
+
     @return: C{True} if a L{recv_stderr} call on this channel would
         immediately return at least one byte; C{False} otherwise.
     @rtype: boolean
-    
+
     @since: 1.1
   */
   recv_stderr_ready : function() {
@@ -4770,10 +4772,10 @@ paramikojs.Channel.prototype = {
     @type nbytes: int
     @return: data.
     @rtype: str
-    
+
     @raise socket.timeout: if no data is ready before the timeout set by
         L{settimeout}.
-    
+
     @since: 1.1
   */
   recv_stderr : function(nbytes) {
@@ -4782,7 +4784,7 @@ paramikojs.Channel.prototype = {
     }
     var out = this.in_stderr_buffer.substring(0, nbytes);
     this.in_stderr_buffer = this.in_stderr_buffer.substring(nbytes);
-        
+
     var ack = this._check_add_window(out.length);
     // no need to hold the channel lock when sending this
     if (ack > 0) {
@@ -4799,11 +4801,11 @@ paramikojs.Channel.prototype = {
   /*
     Returns true if data can be written to this channel without blocking.
     This means the channel is either closed (so any write attempt would
-    return immediately) or there is at least one byte of space in the 
+    return immediately) or there is at least one byte of space in the
     outbound buffer. If there is at least one byte of space in the
     outbound buffer, a L{send} call will succeed immediately and return
     the number of bytes actually written.
-    
+
     @return: C{True} if a L{send} call on this channel would immediately
         succeed or fail
     @rtype: boolean
@@ -4852,15 +4854,15 @@ paramikojs.Channel.prototype = {
     stream is closed.  Applications are responsible for checking that all
     data has been sent: if only some of the data was transmitted, the
     application needs to attempt delivery of the remaining data.
-    
+
     @param s: data to send.
     @type s: str
     @return: number of bytes actually sent.
     @rtype: int
-    
+
     @raise socket.timeout: if no data could be sent before the timeout set
         by L{settimeout}.
-    
+
     @since: 1.1
   */
   send_stderr : function(s) {
@@ -4891,7 +4893,7 @@ paramikojs.Channel.prototype = {
         set by L{settimeout}.
     @raise socket.error: if an error occured before the entire string was
         sent.
-    
+
     @note: If the channel is closed while only part of the data hase been
         sent, there is no way to determine how much data (if any) was sent.
         This is irritating, but identically follows python's API.
@@ -4913,15 +4915,15 @@ paramikojs.Channel.prototype = {
     results.  Unlike L{send_stderr}, this method continues to send data
     from the given string until all data has been sent or an error occurs.
     Nothing is returned.
-    
+
     @param s: data to send to the client as "stderr" output.
     @type s: str
-    
+
     @raise socket.timeout: if sending stalled for longer than the timeout
         set by L{settimeout}.
     @raise socket.error: if an error occured before the entire string was
         sent.
-        
+
     @since: 1.1
   */
   sendall_stderr : function(s) {
@@ -4951,12 +4953,12 @@ paramikojs.Channel.prototype = {
     Return a file-like object associated with this channel's stderr
     stream.   Only channels using L{exec_command} or L{invoke_shell}
     without a pty will ever have data on the stderr stream.
-    
+
     The optional C{mode} and C{bufsize} arguments are interpreted the
     same way as by the built-in C{file()} function in python.  For a
     client, it only makes sense to open this file for reading.  For a
     server, it only makes sense to open this file for writing.
-    
+
     @return: object which can be used for python file I/O.
     @rtype: L{ChannelFile}
 
@@ -4995,7 +4997,7 @@ paramikojs.Channel.prototype = {
     channel will fail instantly.  This is a convenience method, equivalent
     to C{shutdown(0)}, for people who don't make it a habit to
     memorize unix constants from the 1970s.
-    
+
     @since: 1.2
   */
   shutdown_read : function() {
@@ -5008,7 +5010,7 @@ paramikojs.Channel.prototype = {
     channel will fail instantly.  This is a convenience method, equivalent
     to C{shutdown(1)}, for people who don't make it a habit to
     memorize unix constants from the 1970s.
-    
+
     @since: 1.2
   */
   shutdown_write : function() {
@@ -5031,7 +5033,7 @@ paramikojs.Channel.prototype = {
     this.in_window_sofar = 0;
     if(ssh_console.debug) console.debug('Max packet in: ' + max_packet_size + ' bytes');
   },
-    
+
   _set_remote_channel : function(chanid, window_size, max_packet_size) {
     this.remote_chanid = chanid;
     this.out_window_size = window_size;
@@ -5039,7 +5041,7 @@ paramikojs.Channel.prototype = {
     this.active = 1;
     if(ssh_console.debug) console.debug('Max packet out: ' + max_packet_size + ' bytes');
   },
-    
+
   _request_success : function(m) {
     if(ssh_console.debug) console.debug('Sesch channel ' + this.chanid + ' request ok');
   },
@@ -5078,7 +5080,7 @@ paramikojs.Channel.prototype = {
       this.in_stderr_buffer += s;
     }
   },
-    
+
   _window_adjust : function(m) {
     var nbytes = m.get_int();
     if (this.ultra_debug) {
@@ -5862,19 +5864,19 @@ paramikojs.SSHClient.prototype = {
 
 // Source: compress.js
 paramikojs.ZlibCompressor = function () {
-	
+
 };
 
 paramikojs.ZlibCompressor.prototype = {
-	
+
 };
 
 paramikojs.ZlibDecompressor = function () {
-	
+
 };
 
 paramikojs.ZlibDecompressor.prototype = {
-	
+
 };
 
 // Source: dsskey.js
@@ -6389,7 +6391,7 @@ paramikojs.BufferedFile.prototype = {
     @note: If a file is opened in append mode (C{'a'} or C{'a+'}), any seek
         operations will be undone at the next write (as the file position
         will move back to the end of the file).
-    
+
     @param offset: position to move to within the file, relative to
         C{whence}.
     @type offset: int
@@ -7140,7 +7142,7 @@ paramikojs.KexGroup1.prototype = {
 
   _generate_x : function() {
     // generate an "x" (1 < x < q), where q is (p-1)/2.
-    // p is a 128-byte (1024-bit) number, where the first 64 bits are 1. 
+    // p is a 128-byte (1024-bit) number, where the first 64 bits are 1.
     // therefore q can be approximated as a 2^1023.  we drop the subset of
     // potential x where the first 63 bits are 1, because some of those will be
     // larger than q (but this is a tiny tiny subset of potential x).
@@ -7210,7 +7212,7 @@ paramikojs.KexGroup14.prototype = {
   An SSH2 I{Message} is a stream of bytes that encodes some combination of
   strings, integers, bools, and infinite-precision integers (known in python
   as I{long}s).  This class builds or breaks down such a byte stream.
-  
+
   Normally you don't need to deal with anything this low-level, but it's
   exposed for people implementing custom extensions, or features that
   paramiko doesn't support yet.
@@ -7318,7 +7320,7 @@ paramikojs.Message.prototype = {
   get_int : function() {
     return struct.unpack('>I', this.get_bytes(4))[0];
   },
-  
+
   /*
     Fetch a 64-bit int from the stream.
 
@@ -7364,7 +7366,7 @@ paramikojs.Message.prototype = {
 
   /*
     Write bytes to the stream, without any formatting.
-    
+
     @param b: bytes to add
     @type b: str
   */
@@ -7375,7 +7377,7 @@ paramikojs.Message.prototype = {
 
   /*
     Write a single byte to the stream, without any formatting.
-    
+
     @param b: byte to add
     @type b: str
   */
@@ -7386,7 +7388,7 @@ paramikojs.Message.prototype = {
 
   /*
     Add a boolean value to the stream.
-    
+
     @param b: boolean value to add
     @type b: bool
   */
@@ -7401,7 +7403,7 @@ paramikojs.Message.prototype = {
 
   /*
     Add an integer to the stream.
-    
+
     @param n: integer to add
     @type n: int
   */
@@ -7424,7 +7426,7 @@ paramikojs.Message.prototype = {
   /*
     Add a long int to the stream, encoded as an infinite-precision
     integer.  This method only works on positive numbers.
-    
+
     @param z: long int to add
     @type z: long
   */
@@ -7435,7 +7437,7 @@ paramikojs.Message.prototype = {
 
   /*
     Add a string to the stream.
-    
+
     @param s: string to add
     @type s: str
   */
@@ -7449,7 +7451,7 @@ paramikojs.Message.prototype = {
     Add a list of strings to the stream.  They are encoded identically to
     a single string of values separated by commas.  (Yes, really, that's
     how SSH2 does it.)
-    
+
     @param l: list of strings to add
     @type l: list(str)
   */
@@ -7457,7 +7459,7 @@ paramikojs.Message.prototype = {
     this.add_string(l.join(','));
     return this;
   },
-    
+
   _add : function(i) {
     if (typeof i == "string") {
       return this.add_string(i);
@@ -7477,10 +7479,10 @@ paramikojs.Message.prototype = {
   /*
     Add a sequence of items to the stream.  The values are encoded based
     on their type: str, int, bool, list, or long.
-    
+
     @param seq: the sequence of items
     @type seq: sequence
-    
+
     @bug: longs are encoded non-deterministically.  Don't use this method.
   */
   add : function() {
@@ -8051,7 +8053,7 @@ paramikojs.PKey.prototype = {
     object.  If the private key is encrypted and C{password} is not C{None},
     the given password will be used to decrypt the key (otherwise
     L{PasswordRequiredException} is thrown).
-    
+
     @param file_obj: the file to read from
     @type file_obj: file
     @param password: an optional password to use to decrypt the key, if it's
@@ -8059,7 +8061,7 @@ paramikojs.PKey.prototype = {
     @type password: str
     @return: a new key object based on the given private key
     @rtype: L{PKey}
-    
+
     @raise IOError: if there was an error reading the key
     @raise PasswordRequiredException: if the private key file is encrypted,
         and C{password} is C{None}
@@ -8089,12 +8091,12 @@ paramikojs.PKey.prototype = {
   /*
     Write private key contents into a file (or file-like) object.  If the
     password is not C{None}, the key is encrypted before writing.
-    
+
     @param file_obj: the file object to write into
     @type file_obj: file
     @param password: an optional password to use to encrypt the key
     @type password: str
-    
+
     @raise IOError: if there was an error writing to the file
     @raise SSHException: if the key is invalid
   */
@@ -8398,10 +8400,10 @@ paramikojs.PKey.prototype = {
     if(!(window.Components && window.Components.classes)) {
       throw new Error("Unable to write files without Mozilla's Components.classes"); //FIXME
     }
-    
-    var foStream = window.Components.classes["@mozilla.org/network/file-output-stream;1"].createInstance(window.Components.interfaces.nsIFileOutputStream);  
-    foStream.init(f, 0x02 | 0x08 | 0x20, 0600, 0);  
-    var converter = window.Components.classes["@mozilla.org/intl/converter-output-stream;1"].createInstance(window.Components.interfaces.nsIConverterOutputStream);  
+
+    var foStream = window.Components.classes["@mozilla.org/network/file-output-stream;1"].createInstance(window.Components.interfaces.nsIFileOutputStream);
+    foStream.init(f, 0x02 | 0x08 | 0x20, 0600, 0);
+    var converter = window.Components.classes["@mozilla.org/intl/converter-output-stream;1"].createInstance(window.Components.interfaces.nsIConverterOutputStream);
     converter.init(foStream, "UTF-8", 0, 0);
 
     converter.writeString('-----BEGIN ' + tag + ' PRIVATE KEY-----\n');
@@ -8565,7 +8567,7 @@ paramikojs.RSAKey.prototype = {
   write_private_key_file : function(filename, password) {
     this._write_private_key_file('RSA', filename, this._encode_key(), password);
   },
-    
+
   write_private_key : function(file_obj, password) {
     this._write_private_key('RSA', file_obj, this._encode_key(), password);
   },
@@ -10831,7 +10833,7 @@ paramikojs.ssh_exception.NeedRekeyException = function(message) {
   Exception raised when authentication failed for some reason.  It may be
   possible to retry with different credentials.  (Other classes specify more
   specific reasons.)
-  
+
   @since: 1.6
 */
 paramikojs.ssh_exception.AuthenticationException = function(message) {
@@ -10855,12 +10857,12 @@ paramikojs.ssh_exception.PasswordRequiredException = function(message) {
   Exception raised when an authentication type (like password) is used, but
   the server isn't allowing that type.  (It may only allow public-key, for
   example.)
-  
+
   @ivar allowed_types: list of allowed authentication types provided by the
       server (possible values are: C{"none"}, C{"password"}, and
       C{"publickey"}).
   @type allowed_types: list
-  
+
   @since: 1.1
 */
 paramikojs.ssh_exception.BadAuthenticationType = function(message, types) {
@@ -10888,10 +10890,10 @@ paramikojs.ssh_exception.PartialAuthentication = function(types) {
 
 /*
   Exception raised when an attempt to open a new L{Channel} fails.
-  
+
   @ivar code: the error code returned by the server
   @type code: int
-  
+
   @since: 1.6
 */
 paramikojs.ssh_exception.ChannelException = function(code, text) {
@@ -10904,14 +10906,14 @@ paramikojs.ssh_exception.ChannelException = function(code, text) {
 
 /*
   The host key given by the SSH server did not match what we were expecting.
-  
+
   @ivar hostname: the hostname of the SSH server
   @type hostname: str
   @ivar key: the host key presented by the server
   @type key: L{PKey}
   @ivar expected_key: the host key expected
   @type expected_key: L{PKey}
-  
+
   @since: 1.6
 */
 paramikojs.ssh_exception.BadHostKeyException = function(hostname, got_key, expected_key) {
@@ -12648,7 +12650,7 @@ paramikojs.util.setCharAt = function(str, index, ch) {    // how annoying
 
 // Source: win_pageant.js
 paramikojs.win_pageant = function() {
-  
+
 };
 
 paramikojs.win_pageant.prototype = {
@@ -12730,7 +12732,7 @@ paramikojs.win_pageant.prototype = {
     send : function(data) {
       this._response = paramikojs.win_pageant._query_pageant(data);
     },
-    
+
     recv : function(n) {
       if (!this._response) {
         return '';
@@ -12747,4 +12749,7 @@ paramikojs.win_pageant.prototype = {
   }
 };
 
+return paramikojs;
+
+});
