+++ www/js/core/uao/uao_b2u_table.js
@@ -1,3 +1,5 @@
+define([], function () {
+
 var b2uTable = {
 x8140:'\u4E17',
 x8141:'\u4E22',
@@ -19782,3 +19784,7 @@ xFEFC:'\u8262',
 xFEFD:'\u826A',
 xFEFE:'\u8288'
 };
+
+return b2uTable;
+
+});
