+++ www/js/main.js
@@ -1,24 +1,6 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
 (function () {
     "use strict";
-        
+
     function onDeviceReady() {
         // Handle the Cordova pause and resume events
         //document.addEventListener( 'pause', onPause.bind(this), false );
@@ -40,11 +22,11 @@
     function onUnload() {
         //TODO: disconnect
     }
-    
+
     document.addEventListener( 'deviceready', onDeviceReady.bind( this ), false );
     window.addEventListener('load', onLoad.bind(this), false);
     window.addEventListener('unload', onUnload.bind(this), false);
-    
+
     window.app = {
         bbsCore: null
     };
@@ -57,5 +39,29 @@
           return this.siteData[siteName];
         }
     };
+}) ();
 
-} )();
+require.config({
+  paths: {
+    angular: '../lib/angular/angular',
+    "angular-sanitize": '../lib/angular/angular-sanitize',
+    onsen: '../lib/onsen/js/onsenui'
+  },
+  shim: {
+    "angular": {
+      exports: "angular"
+    },
+    "angular-sanitize": ["angular"],
+    "onsen": ["angular"]
+  }
+});
+
+require([
+  'core/sites/ptt/robot',
+  'angular',
+  'frontend/appController',
+  ], function (RobotPtt) {
+  window.RobotPtt = RobotPtt;
+
+  angular.bootstrap(document, ['app']);
+});
